+++ packages/plugin-gallery/web/src/components/gallery-settings-modal.js
@@ -15,7 +15,7 @@ import LayoutSelector from './gallery-controls/layouts-selector';
 import styles from '../../statics/styles/gallery-settings-modal.scss';
 import LayoutControlsSection from './layout-controls-section';
 import { SortableComponent } from './gallery-controls/gallery-items-sortable';
-import { layoutData } from '../../lib/layout-data-provider';
+import { layoutData } from '../layout-data-provider';
 const DIVIDER = 'divider';
 class ManageMediaSection extends Component {
   applyItems = items => {